agent_alpha/requirements.txt
flask
gunicorn
requests
google-genai
google-auth